optimized company list with generic grid service
