+++ src/components/ContactList/ContactList.js
@@ -4,12 +4,12 @@ import AuthService from "../../services/auth.service";
 import { AgGridColumn, AgGridReact } from "ag-grid-react";
 import { ServerSideRowModelModule } from "@ag-grid-enterprise/server-side-row-model";
 
+import GridService from "../../services/grid.service";
+
 import "ag-grid-community/dist/styles/ag-grid.css";
 import "ag-grid-community/dist/styles/ag-theme-alpine.css";
-import axios from "axios";
 import GridTextFilterComponent from "../../shared/components/grid-filters/grid-text-filter.component/grid-text-filter.component";
 
-import { format } from "date-fns";
 import { Button } from "@mui/material";
 import { Link, useNavigate } from "react-router-dom";
 import GridHeaderCheckbox from "../../shared/components/grid-header-checkbox.component";
@@ -26,28 +26,15 @@ const ContactList = () => {
   const onGridReady = (params) => {
     setGridApi(params.api);
     setGridColumnApi(params.columnApi);
-    var datasource = new ServerSideDatasource();
+    var datasource = GridService.ServerSideDatasource("contact/list");
     params.api.setServerSideDatasource(datasource);
   };
 
-  function dateFormatter(params) {
-    if (!params.value) return "";
-    return format(new Date(params.value), "dd/MM/yyyy");
-  }
-
   useEffect(() => {
-    if (!user) {
-      navigate("/auth/login");
-    }
+    if (!user) navigate("/auth/login");
   }, []);
 
-  if (user) {
-    axios.defaults.headers.common["Authorization"] = `Bearer ${user.token}`;
-    axios.defaults.headers.common["token"] = user.token;
-    axios.defaults.headers.common["allowOrigins"] = "*";
-  } else {
-    return <></>;
-  }
+  if (!user) return <></>;
 
   return (
     <div className="container-fluid">
@@ -265,7 +252,7 @@ const ContactList = () => {
             sortable={true}
             filter={true}
             lockPinned={true}
-            valueFormatter={dateFormatter}
+            valueFormatter={GridService.dateFormatter}
             floatingFilterComponent="customTextFloatingFilter"
             floatingFilterComponentParams={{
               suppressFilterButton: true,
@@ -289,7 +276,7 @@ const ContactList = () => {
             sortable={true}
             filter={true}
             lockPinned={true}
-            valueFormatter={dateFormatter}
+            valueFormatter={GridService.dateFormatter}
             floatingFilterComponent="customTextFloatingFilter"
             floatingFilterComponentParams={{
               suppressFilterButton: true,
@@ -309,49 +296,5 @@ const ContactList = () => {
     </div>
   );
 };
-
-function ServerSideDatasource() {
-  return {
-    getRows: function (params) {
-      console.log("[Datasource] - rows requested by grid: ", params.request);
-
-      const sort = params.request.sortModel.map((item) => {
-        return { [`${item.colId}`]: item.sort };
-      });
-      const filters = Object.keys(params.request.filterModel).map((key) => {
-        console.log(key);
-        const item = params.request.filterModel[key];
-        return {
-          logic: "or",
-          filters: [{ field: key, operator: item.type, value: item.filter }],
-        };
-      });
-      const payload = {
-        skip: params.request.startRow,
-        take: 10,
-        group: [],
-        sort,
-        filter: {
-          logic: "and",
-          filters: filters,
-        },
-      };
-
-      axios
-        .post(API_URL + "contact/list", payload)
-        .then(({ data }) => {
-          params.success({
-            rowData: data.result || [],
-            rowCount: data.total,
-          });
-        })
-        .catch((error) => {
-          console.log(error.response);
-          console.log(error.request);
-          params.fail();
-        });
-    },
-  };
-}
 
 export default ContactList;
